toolchain: Add module with raw ELF file definitions
This marks the beginning of our transition to use regular ELF files as
outputs from the assembler and linker.

In addition to the assembler and linker we'll need to implement an
objcopy (I'm quite sure the GNU binutils one will bail out if it doesn't
know about the architecture). objcopy will be necessary for at least
producing the boot ROM image to be embedded into the chipse.

This module currently goes unused, adding it just in case...