rtl-new: Add a simple SDRAM controller and wire it up
The controller as it stands is pretty unoptimized, and wastes 3/4ths
of the 8MiB of SDRAM on-board.

Some optimizations include: keeping rows active between accesses,
adding a cache and doing larger bursts for reads and writes (which
would also help solve the wasted space problem).

The RAM is partially wasted because the board I'm using has DQM{0..3}
all wired low, so the FPGA can't select individual bytes to write.
Currently this is worked around by using whole 32-bit words to store
individual byte.