rtl-new: Make full use of SDRAM capacity
Instead of wasting 3/4ths of every word, allow individual access to
all bytes within the word. For writes, this means we perform a read
command before, because we can't mask off bytes we don't want to
write. This does mean writes are a bit slower now, but it's only by 3
clock cycles (one to issue the read command, two for CAS latency).