rtl: Keep SDRAM rows active for as long as possible
Within each bank we can have one row active at a time, and we don't
need to precharge it between reads/writes. The only times we need to
issue a precharge command is when switching rows, and during
refresh (which precharges rows on all banks).

This has a noticeable improvement on the memory access times.