rtl: Insert combinatorial cuts in some bus subordinates
This helps improve timing in the SoC clock domain. The CSR bus bridge
had a lot of extra delay coming out of the bus, and the internal ram
has a large clk-to-q delay on the BRAM data output, which in
combination with the logic and routing delays went over 10 ns.

The current critical path is inside the MMU, which will be more
problematic to solve.