rtl: Fix sampling MSB of data received over JTAG
It seems that on the iCESugar-Pro, the JTAG adapter has an extra TCK
cycle between Shift-DR and Update-DR where neither happens, which
caused us to miss the final sampled TDI bit, since `update` was raised
one cycle later than expected. The solution is to rely only on
`select` and also shift in on it's negedge.